24-11-22 fixed

+++ 24-11-22/index.js
@@ -1,18 +1,18 @@
 // Funzioni operazioni
 function somma(num1, num2) {
-  return parseInt(num1) + parseInt(num2);
+  return parseFloat(num1) + parseFloat(num2);
 }
 
 function sottrazione(num1, num2) {
-  return parseInt(num1) - parseInt(num2);
+  return parseFloat(num1) - parseFloat(num2);
 }
 
 function divisione(num1, num2) {
-  return parseInt(num1) / parseInt(num2);
+  return parseFloat(num1) / parseFloat(num2);
 }
 
 function moltiplicazione(num1, num2) {
-  return parseInt(num1) * parseInt(num2);
+  return parseFloat(num1) * parseFloat(num2);
 }
 
 const express = require("express");
